Saad Faran Accumulator Circuitry
This is the accumulator logic circuit designed by Muhammad Saad Faran Malik on Logisim. The accumulator is to be connected with the ALU (as shown in the main circuit of the microprocessor). The circuitry was then converted into the form of an IC to minimize the space occupied by it, and the final accumulator is also shown as an IC that is connected with the ALU in the main circuit of the 4-bit microprocessor.