Awais Ahmed Program designed Counter,instruction Register,Instruction Decoder,Rom Circuitry
On each Clock cycle the Program counter starts the count pointing to the next instruction. This is the Program circuit designed by Awais Ahmed on Logisim. The PC is to be connected with the ROM (as shown in the main circuit of the microprocessor). The circuitry was then converted into the form of an IC to minimize the space occupied by it, and the final PC is also shown as an IC that is connected with the Rom in the main circuit of the 4-bit microprocessor.

This is the Instruction register designed by Awais Ahmed on Logisim. The IR is to be connected with the ROM (as shown in the main circuit of the microprocessor). The circuitry was then converted into the form of an IC to minimize the space occupied by it, and the final IR is also shown as an IC that is connected with the Rom in the main circuit of the 4-bit microprocessor.

This is the Instruction decoder designed by Awais Ahmed on Logisim. The ID is to be connected with the ROM (as shown in the main circuit of the microprocessor). The circuitry was then converted into the form of an IC to minimize the space occupied by it, and the final ID is also shown as an IC that is connected with the Rom in the main circuit of the 4-bit microprocessor.

This is the ROM designed by Awais Ahmed on Logisim. The circuitry was then converted into the form of an IC to minimize the space occupied by it, and the final ROM is also shown in the main circuit of the 4-bit microprocessor.